Centralize function application discovery
Keep shared OCI function ownership discovery with the rest of the cleanup discovery logic.
