+++ oci-integration-cleanup/oci_cleanup/domains/extras.py
@@ -125,61 +125,6 @@ class ExtrasMixin:
                 )
         return candidates
 
-    @staticmethod
-    def _is_managed_function(
-        context: CleanupContext, region: str, function_id: str
-    ) -> bool:
-        return any(
-            resource.get("_region") == region
-            and resource_compartment(resource) == context.compartment_id
-            and resource_id(resource) == function_id
-            and function_id.startswith("ocid1.fnfunc.")
-            for resource in context.managed_resources
-        )
-
-    def _owned_function_applications(
-        self, context: CleanupContext, region: str
-    ) -> list[tuple[dict[str, Any], list[dict[str, Any]]]]:
-        applications = self._list_region(
-            region,
-            [
-                "fn",
-                "application",
-                "list",
-                "--compartment-id",
-                context.compartment_id,
-            ],
-        )
-        discovered = []
-        for application in applications:
-            if not exact_owned(
-                application,
-                expected_names={FUNCTION_APP_NAME},
-                compartment_id=context.compartment_id,
-            ):
-                continue
-            functions = self._list_region(
-                region,
-                [
-                    "fn",
-                    "function",
-                    "list",
-                    "--application-id",
-                    resource_id(application),
-                ],
-            )
-            discovered.append(
-                (
-                    application,
-                    [
-                        function
-                        for function in functions
-                        if not is_deleted_or_deleting(function)
-                    ],
-                )
-            )
-        return discovered
-
     def _compute_compartment_ids(self, context: CleanupContext) -> list[str]:
         if self._accessible_compartment_ids is not None:
             return self._accessible_compartment_ids
